Restructured and broke down Programs' layout
