+++ frontend/src/features/activities/programs/utils/formatters.js
@@ -1,0 +1,15 @@
+const DAY_ORDER = [
+  "Monday",
+  "Tuesday",
+  "Wednesday",
+  "Thursday",
+  "Friday",
+  "Saturday",
+  "Sunday",
+];
+
+export const sortDays = (days = []) => {
+  return [...days].sort(
+    (a, b) => DAY_ORDER.indexOf(a) - DAY_ORDER.indexOf(b)
+  );
+};
